+++ src/Moon.Validation/DataValidator.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Reflection;
-using System.Threading.Tasks;
 using Microsoft.Extensions.Localization;
 using Moon.Collections;
 
@@ -32,10 +31,17 @@ namespace Moon.Validation
         /// <summary>
         /// Initializes a new instance of the <see cref="DataValidator" /> class.
         /// </summary>
+        public DataValidator()
+            : this(null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="DataValidator" /> class.
+        /// </summary>
         /// <param name="stringLocalizerFactory">The string localizer factory.</param>
         public DataValidator(IStringLocalizerFactory stringLocalizerFactory)
         {
-            Requires.NotNull(stringLocalizerFactory, nameof(stringLocalizerFactory));
             store = new AttributeStore(stringLocalizerFactory);
         }
 
@@ -43,25 +49,28 @@ namespace Moon.Validation
         /// Returns whether the given object is valid or not.
         /// </summary>
         /// <param name="instance">The object to test. It cannot be null.</param>
-        public async Task<bool> IsValidAsync(object instance)
-        {
-            var results = await ValidateAsync(instance);
-            return !results.Any();
-        }
+        /// <param name="validateRecursively">A value indicating whether to validate all nested objects recursively.</param>
+        public bool IsValid(object instance, bool validateRecursively = false)
+            => !Validate(instance, validateRecursively).Any();
 
         /// <summary>
         /// Validates the given object and returns an enumeration of results for the failures.
         /// </summary>
         /// <param name="instance">The object to test. It cannot be null.</param>
-        public async Task<IEnumerable<ValidationResult>> ValidateAsync(object instance)
+        /// <param name="validateRecursively">A value indicating whether to validate all nested objects recursively.</param>
+        public IEnumerable<ValidationResult> Validate(object instance, bool validateRecursively = false)
         {
             Requires.NotNull(instance, nameof(instance));
 
+            return Validate(CreateObjectContext(instance), validateRecursively);
+        }
+
+        private IEnumerable<ValidationResult> Validate(ValidationContext objectContext, bool validateRecursively)
+        {
             var results = new List<ValidationResult>();
-            var objectContext = CreateObjectContext(instance);
 
             // Step 1: Validate the object properties' validation attributes
-            results.AddRange(ValidateProperties(objectContext));
+            results.AddRange(ValidateProperties(objectContext, validateRecursively));
 
             if (results.Any())
             {
@@ -76,36 +85,40 @@ namespace Moon.Validation
                 return results;
             }
 
-            // Step 3: Execute IAsyncValidatableObject.ValidateAsync implementation
-            var asyncValidatable = instance as IAsyncValidatableObject;
-
-            if (asyncValidatable != null)
-            {
-                var asyncResults = await asyncValidatable.ValidateAsync(objectContext);
-                results.AddRange(asyncResults.Where(x => x != ValidationResult.Success));
-            }
-
-            // Step 4: Execute IValidatableObject.Validate implementation
-            var validatable = instance as IValidatableObject;
+            // Step 3: Execute IValidatableObject.Validate implementation
+            var validatable = objectContext.ObjectInstance as IValidatableObject;
 
             if (validatable != null)
             {
-                var syncResults = validatable.Validate(objectContext);
-                results.AddRange(syncResults.Where(x => x != ValidationResult.Success));
+                results.AddRange(validatable.Validate(objectContext)
+                    .Select(x => new ValidationResult(x.ErrorMessage, AppendMemberNames(objectContext, x.MemberNames)))
+                    .Where(x => x != ValidationResult.Success));
             }
 
             return results;
         }
 
-        private IEnumerable<ValidationResult> ValidateProperties(ValidationContext objectContext)
+        private IEnumerable<ValidationResult> ValidateProperties(ValidationContext objectContext, bool validateRecursively)
         {
             var results = new List<ValidationResult>();
-            var propsToValidate = GetPropertiesToValidate(objectContext);
+            var propsToValidate = GetPropertiesToValidate(objectContext, validateRecursively);
 
             foreach (var property in propsToValidate)
             {
+                var propertyResults = new List<ValidationResult>();
                 var attributes = store.GetPropertyValidationAttributes(property.Context);
-                results.AddRange(GetResults(property.Value, property.Context, attributes));
+
+                propertyResults.AddRange(GetResults(property.Value, property.Context, attributes));
+
+                if (propertyResults.Count == 0 && property.Value != null && validateRecursively)
+                {
+                    var memberName = property.Context.MemberName;
+                    var propertyObjectContext = CreatePropertyObjectContext(objectContext, memberName, property.Value);
+
+                    propertyResults.AddRange(Validate(propertyObjectContext, true));
+                }
+
+                results.AddRange(propertyResults);
             }
 
             return results;
@@ -121,7 +134,7 @@ namespace Moon.Validation
             return results;
         }
 
-        private IEnumerable<PropertyToValidate> GetPropertiesToValidate(ValidationContext objectContext)
+        private IEnumerable<PropertyToValidate> GetPropertiesToValidate(ValidationContext objectContext, bool validateRecursively)
         {
             var properties = objectContext.ObjectType.GetRuntimeProperties()
                 .Where(p => IsPublic(p) && p.GetIndexParameters().Empty());
@@ -130,7 +143,7 @@ namespace Moon.Validation
             {
                 var context = CreatePropertyContext(objectContext, property.Name);
 
-                if (store.GetPropertyValidationAttributes(context).Any())
+                if (validateRecursively || store.GetPropertyValidationAttributes(context).Any())
                 {
                     yield return new PropertyToValidate {
                         Context = context,
@@ -175,12 +188,26 @@ namespace Moon.Validation
             return context;
         }
 
+        private ValidationContext CreatePropertyObjectContext(ValidationContext objectContext, string memberName, object value)
+        {
+            var context = new ValidationContext(value, objectContext.Items) { MemberName = memberName };
+            context.DisplayName = store.GetTypeDisplayName(context);
+            return context;
+        }
+
         private ValidationContext CreatePropertyContext(ValidationContext objectContext, string propertyName)
         {
-            var context = new ValidationContext(objectContext.ObjectInstance, objectContext.Items) { MemberName = propertyName };
+            var memberName = AppendMemberName(objectContext, propertyName);
+            var context = new ValidationContext(objectContext.ObjectInstance, objectContext.Items) { MemberName = memberName };
             context.DisplayName = store.GetPropertyDisplayName(context);
             return context;
         }
+
+        private IEnumerable<string> AppendMemberNames(ValidationContext objectContext, IEnumerable<string> memberNames)
+            => memberNames.Select(n => AppendMemberName(objectContext, n));
+
+        private string AppendMemberName(ValidationContext objectContext, string memberName)
+            => objectContext.MemberName != null ? $"{objectContext.MemberName}.{memberName}" : memberName;
 
         private bool TryValidate(object value, ValidationContext validationContext, ValidationAttribute attribute, out ValidationResult result)
         {